Add a new 'catalogue' FPE, which takes font paths from symlinks in a dir.
This patch adds a new FPE type, which will match font path elements of the
form

	catalogue:<dir>

The dir specified after the catalogue: prefix will be scanned for symlinks
and each symlink destination will be added as a local fontfile FPE.
The symlink can be suffixed by attributes, such as 'unscaled', which
will be passed through to the underlying fontfile FPE.  Except the new
attribute 'pri' which will be used for ordering the fontfile FPEs.

An example configuration:

	75dpi:unscaled:pri=20 -> /usr/share/X11/fonts/75dpi
	ghostscript:pri=60 -> /usr/share/fonts/default/ghostscript
	misc:unscaled:pri=10 -> /usr/share/X11/fonts/misc
	type1:pri=40 -> /usr/share/X11/fonts/Type1
	type1:pri=50 -> /usr/share/fonts/default/Type1

will add /usr/share/X11/fonts/misc as the first FPE with the attribute
'unscaled', second FPE will be /usr/share/X11/fonts/75dpi, also with
the attribute unscaled etc.  This is functionally equivalent to setting
the following font path:

	/usr/share/X11/fonts/misc:unscaled,
	/usr/share/X11/fonts/75dpi:unscaled,
	/usr/share/X11/fonts/Type1,
	/usr/share/fonts/default/Type1,
	/usr/share/fonts/default/ghostscript

The motivation is to let font packages add a symlink to the new font
directory they provide instead of rewriting either the Xorg config file
or the xfs config file.