Always scan catalogue dirs at startup, even if their mtime < 0
Works around bug in VMWare that caused the Solaris kernel to fail to read
the time-of-day chip on boot, and start with a time < 0 (i.e. back to the
1960's) - when the system was then installed from this mode, Xorg wouldn't
start after reboot, since the mtime would be < 0 and the catalogue dir was
then skipped as not changed since reading at the 0 initially set in the
cat->mtime.

Fixes OpenSolaris bug #4780
   <http://defect.opensolaris.org/bz/show_bug.cgi?id=4780>