config: replace obsolete AM_CONFIG_HEADER with AC_CONFIG_HEADERS
There are 2 headers to generate. The first one (config.h)
is picked up by autoheader to generate the familiar config.h.in
input file. The others in the list (or in subsequent AC_CONFIG_HEADERS
macro calls) are generated from their existing matching template,
e.g. fontconf.h.in.

When multiple headers are listed in the same macro call,
they cannot be separated by a new line like we do in AC_OUTPUT.

Reviewed-by: Dan Nicholson <dbn.lists@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>