Build fix for platforms which don't have weak linkage
Since we fix this by removing the serverGeneration symbol, assuming
an external definition will be provided, this means on Windows libXfont
can only be built as a static library (since PE shared libraries cannot
contain undefined symbols).  This produces a libXfont which might only
be useful to the xserver, but the only other users we might care about
are xfs, which is obsolete, and bdftopcf, which fortunately doesn't
pull in any objects which reference serverGeneration from libXfont.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Reviewed-by: Colin Harrison <colin.harrison@virgin.net>